design: 메인페이지 레이아웃, 로그인페이지 레이아웃 추가
