--- src/components/LayoutContent.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from "react-router-dom";
-
-export default function LayoutContent() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-}
