+++ src/components/LayoutMain.Content.tsx
@@ -1,0 +1,7 @@
+export default function Content({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="pt-14">
+      <div className="p-10">{children}</div>
+    </div>
+  );
+}
