+++ src/index.css
@@ -1,3 +1,12 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+input {
+  border-bottom: 1px solid black;
+}
+
+button {
+  border: 1px solid black;
+  padding: 0 1rem;
+}
