+++ src/routes/login.tsx
@@ -1,28 +1,5 @@
-import { useNavigate } from "react-router-dom";
-import { storage } from "../utils/storage";
+import FormLogin from "../components/FormLogin";
 
 export default function Login() {
-  const navigator = useNavigate();
-
-  return (
-    <>
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          storage.set("ACCESS_TOKEN", "temp");
-          navigator("/");
-        }}
-      >
-        <label>
-          id
-          <input type="email" name="id" />
-        </label>
-        <label>
-          password
-          <input type="password" name="password" />
-        </label>
-        <button type="submit">login</button>
-      </form>
-    </>
-  );
+  return <FormLogin />;
 }
