+++ tailwind.config.cjs
@@ -2,7 +2,9 @@
 module.exports = {
   content: ["./index.html", "./src/**/*.{js,jsx,ts,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      haederHeight: "3.5rem",
+    },
   },
   plugins: [],
 };
